in the process of swithing from NS timers to SKActions

+++ basicBattleAIPoC/Models/Ally.swift
@@ -11,10 +11,10 @@ import Foundation
 class Ally {
     var health: Int
     var attackDamage: Int
-    var attackSpeed: Int
+    var attackSpeed: Double
     var stance: Bool
     
-    init(health: Int, attackDamage: Int, attackSpeed: Int, stance: Bool) {
+    init(health: Int, attackDamage: Int, attackSpeed: Double, stance: Bool) {
         self.health = health
         self.attackDamage = attackDamage
         self.attackSpeed = attackSpeed
